requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.0
channels==4.0.0
channels-redis==4.1.0
redis==5.0.1
psycopg2-binary==2.9.9
daphne==4.0.0
python-dotenv==1.0.0
django-redis==5.4.0
django-ratelimit==4.1.0
pytest==7.4.3
pytest-django==4.7.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0
gunicorn==21.2.0
whitenoise==6.6.0 